--- lib/loading_screen.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class LoadingScreen extends StatelessWidget {
-  const LoadingScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold();
-  }
-}
